+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - favicon configuration to environments and apps
 - `validationIcon` configuration to environments
 - `maxFileSize` to `FormElement`
+- `restrictFileSize` to `FormElement`
 
 ## [11.0.6] - 2025-06-16